Refactor css and components structure
