+++ src/components/FileInput/FileInput.module.css
@@ -1,0 +1,24 @@
+.dropzoneWrapper {
+    display: flex;
+    justify-content: center;
+    padding-bottom: 2em;
+}
+
+.dropzoneFile {
+    border: 1px dashed #8997A5;
+    border-radius: 5px;
+    display: flex;
+    cursor: pointer;
+}
+
+.dropzoneText {
+    width: 100%;
+    text-align: center;
+    color: #091440;
+    font-size: 13px;
+}
+
+.dropzoneAcceptStyle {
+    border: 1px solid #2970FF;
+    background: #2970FF;
+}
