+++ src/components/FileInput/FileInputLine.module.css
@@ -1,0 +1,19 @@
+.fileInputLine {
+    display: flex;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    height: 50px;
+    width: 100%;
+}
+
+.textFileName {
+    margin-top: 5px;
+    margin-left: 5px;
+    font-size: 13px;
+    color: #48566E;
+    width: 100%;
+}
+
+.deleteIcon {
+    cursor: pointer;
+}
