+++ src/components/HashInput/HashInput.module.css
@@ -1,0 +1,9 @@
+.hashInputWrapper {
+    margin-top: 57px;
+    margin-bottom: 40px;
+}
+
+.addButtonText {
+    color: #2970FF;
+    font-size: 13px;
+}
