+++ src/components/ModalNewDigestInfo.jsx
@@ -1,16 +1,15 @@
 import React from "react";
 import { Modal, H1, Icon } from "pi-ui";
-import styles from "./components.module.css";
 
 const ModalNewDigestInfo = ({ show, onClose }) => (
   <Modal show={show} contentStyle={{ padding: "1rem" }} onClose={onClose}>
-    <div className={styles.container}>
-      <div className={styles.iconRow}>
+    <div className="modalContainer">
+      <div className="modalIconRow">
         <Icon type="info" size="xlg" />
       </div>
       <div>
         <H1>Create new digest</H1>
-        <div className={styles.infoText}>
+        <div className="modalInfoText">
           The timestamp service allows you to create a Proof-of-Existence of a
           given file. A digital signature (digest) of each file is calculated
           and sent to the Dcrtime server which hourly will calculate the merkle
