+++ src/components/TimestampForm/TimestampForm.module.css
@@ -1,0 +1,30 @@
+/* TimestampForm styles */
+
+.card {
+    background: #FFFFFF 0% 0% no-repeat padding-box;
+    box-shadow: 0px 4px 16px #48566E36;
+    opacity: 1;
+    width: 100%;
+    position: relative;
+    margin-top: 50px;
+}
+
+.content {
+    display: flex;
+    flex-direction: column;
+    margin-right: 30px;
+    margin-left: 30px;
+}
+
+.submitButtonWrapper {
+    display: flex;
+    justify-content: flex-end;
+    margin-top: 20px;
+}
+
+.submitButton {
+    font-size: 13px;
+    width: 125px;
+    height: 44px;
+    box-shadow: 0px 5px 13px #3D587333;;
+}
